+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updates to terraform sample deployment.
+- Fixed AzureDevops Environment Variable Capitalization Issue
 
 ----------------
 ## [1.0.4] - 2019-06-11